# [refactor] #3273: Move to tokio actors in p2p::peer