--- a/Cargo.toml
+++ b/Cargo.toml
@@ -16,8 +16,6 @@ categories = ["cryptography::cryptocurrencies"]
 [workspace]
 resolver = "2"
 members = [
-  "actor",
-  "actor/derive",
   "cli",
   "cli/derive",
   "client",