--- a/actor/derive/tests/ui_pass/generics.rs
+++ b/actor/derive/tests/ui_pass/generics.rs
@@ -1,7 +1,0 @@
-trait Predicate {}
-
-#[derive(iroha_actor::Message)]
-#[message(result = "i32")]
-struct M<A: Predicate>(A);
-
-fn main() {}