--- a/actor/derive/tests/ui_pass/simple.rs
+++ b/actor/derive/tests/ui_pass/simple.rs
@@ -1,3 +1,0 @@
-#[derive(iroha_actor::Message)]
-struct M;
-fn main() {}