--- a/actor/derive/tests/ui_pass/simple_result.rs
+++ b/actor/derive/tests/ui_pass/simple_result.rs
@@ -1,4 +1,0 @@
-#[derive(iroha_actor::Message)]
-#[message(result = "i32")]
-struct M;
-fn main() {}