--- a/cli/Cargo.toml
+++ b/cli/Cargo.toml
@@ -43,7 +43,6 @@ iroha_core = { version = "=2.0.0-pre-rc.13", path = "../core" }
 iroha_macro = { version = "=2.0.0-pre-rc.13", path = "../macro" }
 iroha_logger = { version = "=2.0.0-pre-rc.13", path = "../logger" }
 iroha_futures = { version = "=2.0.0-pre-rc.13", path = "../futures" }
-iroha_actor = { version = "=2.0.0-pre-rc.13", path = "../actor" }
 iroha_data_model = { version = "=2.0.0-pre-rc.13", path = "../data_model", features = ["http"] }
 iroha_telemetry = { version = "=2.0.0-pre-rc.13", path = "../telemetry", optional = true }
 iroha_version = { version = "=2.0.0-pre-rc.13", path = "../version", features = ["http"] }