--- a/core/src/lib.rs
+++ b/core/src/lib.rs
@@ -27,7 +27,7 @@ use crate::{
 pub const TX_RETRIEVAL_INTERVAL: Duration = Duration::from_millis(100);
 
 /// Specialized type of Iroha Network
-pub type IrohaNetwork = iroha_p2p::Network<NetworkMessage>;
+pub type IrohaNetwork = iroha_p2p::NetworkHandle<NetworkMessage>;
 
 /// Ids of peers.
 pub type PeersIds = DashSet<<Peer as Identifiable>::Id>;
@@ -52,7 +52,7 @@ pub type PermissionTokenDefinitionsMap =
 pub type EventsSender = broadcast::Sender<Event>;
 
 /// The network message
-#[derive(Clone, Debug, Encode, Decode, iroha_actor::Message)]
+#[derive(Clone, Debug, Encode, Decode)]
 pub enum NetworkMessage {
     /// Blockchain message
     SumeragiPacket(Box<SumeragiPacket>),