@SamHSmith 2023-04-03
I think this is clearly an improvement. Even if you don't even look at the code. The line count went down by 500. And you improved TPS by roughly 16 %, that's very good. I think it is obvious that we should merge this.